// src/types.ts
export interface PersonalInfoFormData {
  name: string;
  email: string;
  phoneNumber: string;
  address: string;
}

export interface ServicesDetailsFormData {
  vehicleType: string;
  vehicleModelNo: string;
}

export interface BookingDetailsFormData {
  appointmentDate: string;
  appointmentTime: string;
}